LaFbWXIf - Refactor OIDC client requests
- Move the OIDC client requests into its own resource which is just responsible for behaving as an OIDC client. This is required for when a user needs to authenticate themself to be able to register their client software.
- Move the OIDC client ID and secret into config.
