+++ oidc-provider.yml
@@ -13,3 +13,7 @@ database:
   url: ${DATABASE_URL:-jdbc:postgresql://localhost:15432/store}
   user: user
   password: pass
+
+clientId: admin_client_id
+
+clientSecret: admin_client_secret
