+++ src/main/java/uk/gov/di/OidcProviderApplication.java
@@ -21,6 +21,7 @@ import uk.gov.di.resources.AuthorisationResource;
 import uk.gov.di.resources.ClientRegistrationResource;
 import uk.gov.di.resources.LoginResource;
 import uk.gov.di.resources.LogoutResource;
+import uk.gov.di.resources.OidcClientResource;
 import uk.gov.di.resources.RegistrationResource;
 import uk.gov.di.resources.TokenResource;
 import uk.gov.di.resources.UserInfoResource;
@@ -73,6 +74,7 @@ public class OidcProviderApplication extends Application<OidcProviderConfigurati
         var authenticationService = getAuthenticationService(configuration);
         var tokenService = new TokenService(configuration);
 
+        env.jersey().register(new OidcClientResource(configuration));
         env.jersey().register(new AuthorisationResource(clientService));
         env.jersey().register(new LoginResource(authenticationService));
         env.jersey().register(new RegistrationResource(authenticationService));
@@ -81,7 +83,7 @@ public class OidcProviderApplication extends Application<OidcProviderConfigurati
                 .register(new TokenResource(tokenService, clientService, authorizationCodeService));
         env.jersey().register(new LogoutResource());
         env.jersey().register(new RegistrationResource(authenticationService));
-        env.jersey().register(new ClientRegistrationResource(clientService));
+        env.jersey().register(new ClientRegistrationResource(clientService, configuration));
         env.jersey()
                 .property(ServerProperties.LOCATION_HEADER_RELATIVE_URI_RESOLUTION_DISABLED, true);
     }
